refactor: merge /api/chat-stream to /api/openai
