+++ app/requests.ts
@@ -161,7 +161,7 @@ export async function requestChatStream(
   const reqTimeoutId = setTimeout(() => controller.abort(), TIME_OUT_MS);
 
   try {
-    const res = await fetch("/api/chat-stream", {
+    const res = await fetch("/api/openai", {
       method: "POST",
       headers: {
         "Content-Type": "application/json",
